+++ Server/src/DTO/create-student.dto.ts
@@ -1,6 +1,3 @@
-
-
-
 export class Createstudentdto {
   readonly name: string;
   readonly Class: string;
